Moved ChangeEmailRequests model class out of model.py
